Create README.md
Here is an example of a  task , that our program can solve for us:
There are two alloys of copper and tin. The first alloy contains 40% copper, and the second alloy -60% copper. How much does it take for each alloy to get 10 kg of a new alloy that would contain 54% copper.
This program  allows us to solve this kind of tasks. User just should enter a percentage of the first and the second alloys and also, the percentage and the mass of final alloy and it counts what masses of each we should take. Before we start to write our program let’s do some calculations in order to solve this task.
first_coef = 60 / 100 = 0.6
second_coef = 40 / 100 = 0.4
third_coef = 54 * 10 / 100 = 5.4
0.6 *X + 0.4 * (10 – X) = 5.4
0.6 * X + 0.4*10 – 0.4 * X = 5.4
0.6 * X – 0.4 * X = 5.4 – 0.4 *10
X * (0.6 – 0.4) = 5.4 – 0.4 * 10 
X = 5.4 – 0.4 * 10 /  ( 0.6 – 0.4)
Now we clearly see how we can implement our program.